app/requirements.txt
flask
python-dotenv
bson
mongoengine
flask-jwt-extended
flask_restplus
Werkzeug
mailjet_rest
flask-cors